更新【edonet】index 页面

+++ webpack/utils/rmdir.js
@@ -2,38 +2,35 @@
 
 const
     fs = require('fs'),
-    path = require('path');
+    path = require('path'),
+    stat = require('./stat'),
+    readdir = require('./readdir');
 
 
-// 异步移除目录
-function rmdir(dir, callback) {
-    fs.stat(dir, function (err, stats) {
+// 异步删除目录
+function rmdirAsync (dir, callback) {
+    stat(dir, function (err, stats) {
         if (err) {
             return callback(null);
         }
 
         if (stats.isDirectory()) {
-            fs.readdir(dir, function (err, files) {
+            readdir(dir, function (err, files) {
                 if (err) {
                     return callback(err);
                 }
 
-                let len = files.length,
-                    count = 1;
-
-                if (len) {
-                    files.forEach(function (file) {
-                        rmdir(path.join(dir, file), function (err) {
-                            if (err) {
-                                return callback(err);
-                            }
-
-                            count ++ === len && fs.rmdir(dir, callback);
+                Promise
+                    .all(files.map(function (file) {
+                        return new Promise(function (resolve, reject) {
+                            rmdirAsync(path.join(dir, file), function (err) {
+                                err ? reject(err) : resolve();
+                            });
                         });
-                    });
-                } else {
-                    fs.rmdir(dir, callback);
-                }
+                    }))
+                    .then(function () {
+                        fs.rmdir(dir, callback);
+                    }, callback);
             });
         } else {
             fs.unlink(dir, callback);
@@ -41,36 +38,51 @@ function rmdir(dir, callback) {
     });
 }
 
-// 同步移除目录
-function rmdirSync(dir) {
-    let stats, files;
 
-    try {
-        stats = fs.statSync(dir);
-    } catch (e) {
+// 同步删除目录
+function rmdirSync (dir) {
+    let stats = stat(dir);
+
+    if (!stats) {
         return true;
     }
 
-    try {
-        if (stats.isDirectory()) {
-            files = fs.readdirSync(dir);
+    if (stats.isDirectory()) {
+        let files = readdir(dir);
 
-            files.forEach(function (file) {
-                rmdirSync(path.join(dir, file));
-            });
-
-            fs.rmdirSync(dir);
-        } else {
-            fs.unlinkSync(dir);
+        if (!files) {
+            return false;
         }
 
+        for (let file of files) {
+            if (rmdirSync(path.join(dir, file)) === false) {
+                return false;
+            }
+        }
+
+        return trydo(function () {
+            fs.rmdirSync(dir);
+        });
+    } else {
+        return trydo(function () {
+            fs.unlinkSync(dir);
+        });
+    }
+}
+
+
+// 过滤异常
+function trydo(handler) {
+    try {
+        handler();
         return true;
-    } catch (e) {
+    } catch(e) {
         return false;
     }
 }
 
 // 抛出接口
 module.exports = function (dir, callback) {
-    callback ? rmdir(dir, callback) : rmdirSync(dir);
+    return typeof callback === 'function' ?
+        rmdirAsync(dir, callback) : rmdirSync(dir);
 };
